requirements.txt
asgiref==3.8.1
astroid==3.1.0
attrs==23.2.0
<<<<<<< HEAD
Django>=4.2
=======
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
dill==0.3.8
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==5.0.4
django-allauth==0.52.0
>>>>>>> origin/back_silvina
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-spectacular==0.27.2
drf-standardized-errors==0.13.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
isort==5.13.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
platformdirs==4.2.1
oauthlib==3.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pylint==3.1.0
pyOpenSSL==24.1.0
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tomlkit==0.12.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
zope.interface==6.3
urllib3==2.2.1
